+++ src/placePromos.js
@@ -13,7 +13,7 @@ import sortBy from 'fkit/dist/sortBy'
 import UAParser from 'ua-parser-js'
 
 import match from './match'
-import { timestamp } from './utils'
+import { age, timestamp } from './utils'
 
 const PROMO_PROPERTIES = ['creativeId', 'promoId', 'slotId', 'groupId', 'campaignId']
 
@@ -38,6 +38,7 @@ export default function placePromos (promos, user, window) {
   // The client state object used to match promo constraints. This object
   // should contain everything we want to match against.
   const clientState = {
+    age,
     browser: parser.getBrowser(),
     device: parser.getDevice(),
     os: parser.getOS(),